# json_string_unescape: assume the string doesn't need escaping